requirements.txt
requests
beautifulsoup4
DrissionPage
wxpusher
schedule
lxml
pillow
